requirements.txt
pyspark==3.5.0
streamlit
newsapi-python
pandas
